# openapi.yaml
openapi: 3.0.3
info:
  title: Midori API
  version: 0.1.0
  description: >
    OpenAPI specification สำหรับระบบ Midori - AI-powered website generation platform.
    รองรับการจัดการ authentication, projects, templates, conversations, และ deployment.
  contact:
    name: Midori Team
    url: https://midori.local
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.midori.production
    description: Production server (ถ้ามี)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: midori-session
      description: Session cookie สำหรับ authentication
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: เกิดข้อผิดพลาด
        details:
          type: object
          additionalProperties: true
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [private, public]
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: เข้าสู่ระบบ
      description: Login ด้วย email และ password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: เข้าสู่ระบบสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: เข้าสู่ระบบสำเร็จ
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: ข้อมูลไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Email หรือ password ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: สมัครสมาชิก
      description: สร้างบัญชีผู้ใช้ใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                displayName:
                  type: string
                  minLength: 2
      responses:
        '200':
          description: สมัครสมาชิกสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: ข้อมูลไม่ถูกต้องหรือ email ซ้ำ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/validate:
    get:
      tags: [Authentication]
      summary: ตรวจสอบ session validity
      description: ใช้สำหรับตรวจสอบว่า session ยังใช้งานได้อยู่หรือไม่
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Session ยังใช้งานได้
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Session ไม่ถูกต้องหรือหมดอายุ
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: false
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
    post:
      tags: [Authentication]
      summary: ตรวจสอบ session (POST)
      description: รองรับ POST method สำหรับ middleware
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Session ยังใช้งานได้
        '401':
          description: Session ไม่ถูกต้อง

  # ========== Projects ==========
  /api/projects:
    get:
      tags: [Projects]
      summary: ดึงรายการโปรเจกต์ทั้งหมด
      description: ดึงรายการโปรเจกต์เรียงตามวันที่สร้างล่าสุด
      responses:
        '200':
          description: รายการโปรเจกต์
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '500':
          description: เกิดข้อผิดพลาด
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Projects]
      summary: สร้างโปรเจกต์ใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name:
                  type: string
                  example: My Website
                description:
                  type: string
                  example: A beautiful website
                visibility:
                  type: string
                  enum: [private, public]
                  default: private
      responses:
        '200':
          description: สร้างโปรเจกต์สำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: ข้อมูลไม่ครบถ้วน
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/visual-edit/upload-image:
    post:
      tags: [Visual Edit]
      summary: Upload รูปภาพสำหรับ visual editing
      description: Upload รูปไปยัง Cloudflare R2 และบันทึก metadata
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, projectId, blockId, field]
              properties:
                file:
                  type: string
                  format: binary
                  description: ไฟล์รูปภาพ (JPEG, PNG, GIF, WEBP) ไม่เกิน 10MB
                projectId:
                  type: string
                  format: uuid
                blockId:
                  type: string
                field:
                  type: string
                compressed:
                  type: boolean
                  default: false
                originalSize:
                  type: integer
                  description: ขนาดไฟล์เดิมก่อน compress (bytes)
      responses:
        '200':
          description: Upload สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  url:
                    type: string
                    format: uri
                  imageAssetId:
                    type: string
                    format: uuid
                  meta:
                    type: object
                    properties:
                      filename:
                        type: string
                      size:
                        type: integer
                      type:
                        type: string
                      uploadedAt:
                        type: string
                        format: date-time
        '400':
          description: ข้อมูลไม่ครบหรือไฟล์ไม่ถูกต้อง
        '500':
          description: Upload ล้มเหลว
...
  /api/preview/daytona:
    post:
      tags: [Preview]
      summary: สร้าง Daytona sandbox ใหม่
      description: สร้าง sandbox สำหรับ preview โปรเจกต์
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, files]
              properties:
                projectId:
                  type: string
                  format: uuid
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        type: string
                      content:
                        type: string
      responses:
        '200':
          description: สร้าง sandbox สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId:
                    type: string
                  url:
                    type: string
                    format: uri
                  state:
                    type: string

    get:
      tags: [Preview]
      summary: ดึงสถานะ sandbox
      description: ตรวจสอบสถานะและ heartbeat ของ sandbox
      parameters:
        - name: sandboxId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: สถานะ sandbox
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId:
                    type: string
                  state:
                    type: string
                  url:
                    type: string
        '400':
          description: ไม่มี sandboxId
        '404':
          description: ไม่พบ sandbox

    put:
      tags: [Preview]
      summary: อัปเดตไฟล์ใน sandbox
      parameters:
        - name: sandboxId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        type: string
                      content:
                        type: string
      responses:
        '200':
          description: อัปเดตสำเร็จ
        '400':
          description: ไม่มี sandboxId
        '404':
          description: ไม่พบ sandbox

    delete:
      tags: [Preview]
      summary: ลบ sandbox
      parameters:
        - name: sandboxId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ลบสำเร็จ
        '400':
          description: ไม่มี sandboxId
        '500':
          description: ลบล้มเหลว